move : Sell Repository 분리
