+++ src/main/java/bc1/gream/domain/sell/repository/SellRepository.java
@@ -1,6 +1,6 @@
-package bc1.gream.domain.order.repository;
+package bc1.gream.domain.sell.repository;
 
-import bc1.gream.domain.order.entity.Sell;
+import bc1.gream.domain.sell.entity.Sell;
 import bc1.gream.domain.user.entity.User;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
