+++ src/main/java/bc1/gream/domain/sell/repository/SellRepositoryCustom.java
@@ -1,6 +1,6 @@
-package bc1.gream.domain.order.repository;
+package bc1.gream.domain.sell.repository;
 
-import bc1.gream.domain.order.entity.Sell;
+import bc1.gream.domain.sell.entity.Sell;
 import bc1.gream.domain.product.entity.Product;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
