+++ src/test/java/bc1/gream/domain/sell/repository/SellRepositoryCustomImplTest.java
@@ -4,9 +4,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import bc1.gream.domain.gifticon.repository.GifticonRepository;
 import bc1.gream.domain.order.entity.Gifticon;
-import bc1.gream.domain.order.entity.Sell;
-import bc1.gream.domain.order.repository.SellRepository;
-import bc1.gream.domain.order.repository.SellRepositoryCustomImpl;
+import bc1.gream.domain.sell.entity.Sell;
 import bc1.gream.domain.product.entity.Product;
 import bc1.gream.domain.product.repository.ProductRepository;
 import bc1.gream.domain.user.entity.User;
